requirements.txt
google-generativeai
python-dotenv
colorama
google-adk
